# apps/api/openapi/public-api-v2.generated.json
{
  "components": {
    "responses": {
      "Conflict": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "The Agent/session state rejects this action, or an Idempotency-Key is already processing or was reused for a different request."
      },
      "Forbidden": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "The caller cannot consume this Agent."
      },
      "InternalError": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "The request failed unexpectedly."
      },
      "InvalidRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "The request shape or query value is invalid."
      },
      "NotFound": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "The resource was not found for this caller."
      },
      "RateLimited": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "The caller token exceeded the public API request budget for the current window.",
        "headers": {
          "Retry-After": {
            "description": "Seconds to wait before retrying the request.",
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          }
        }
      },
      "Unauthenticated": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "A valid Access Token is required."
      }
    },
    "schemas": {
...
      "PublicFileResponse": {
        "additionalProperties": false,
        "description": "A single public file.",
        "properties": {
          "file": {
            "$ref": "#/components/schemas/PublicFile",
            "description": "Public file metadata."
          }
        },
        "required": ["file"],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "accessToken": {
        "bearerFormat": "mosoo Access Token",
        "description": "Use Authorization: Bearer msp_... . Project API keys can configure Agents, execute Sessions, and access files only in their Project. Key and account management require console or CLI login. Legacy account tokens are rejected.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Project-scoped durable Threads for private saved Agents. The existing Thread ID remains the conversation handle for reading, continuing, cancelling, and downloading files, regardless of creation channel or current publication state. v1 retains its existing admission and published-configuration behavior.",
    "title": "mosoo Public Thread API",
    "version": "v2"
  },
  "openapi": "3.1.0",
  "paths": {
    "/agents/{agentId}/files": {
      "post": {
        "description": "Uploads a file into the Agent API Endpoint's Project draft scope before a Thread exists. Use the returned file ID in create-thread or send-events resources.",
        "parameters": [
          {
            "description": "Agent API Endpoint ID from the Agent's API Access panel. v1 IDs are bare ULIDs.",
            "example": "01J00000000000000000000001",
            "in": "path",
            "name": "agentId",
            "required": true,
            "schema": {
              "example": "01J00000000000000000000001",
              "format": "ulid",
              "pattern": "^[0-7][0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{25}$",
              "type": "string",
              "x-default": "01J00000000000000000000001"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "file": {
                    "description": "File bytes to upload before attaching them to a Thread.",
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": ["file"],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Upload an Agent file",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "example": {
                  "file": {
                    "createdAt": "2026-05-19T00:02:00.000Z",
                    "id": "01J0000000000000000000000J",
                    "mimeType": "text/plain",
                    "name": "brief.txt",
                    "size": 19
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/PublicFileResponse"
                }
              }
            },
            "description": "Uploaded file."
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/files/{fileId}/content": {
      "get": {
        "description": "Downloads bytes for a ready Thread attachment or Agent artifact. The file must belong to a public Thread visible to the Access Token caller.",
        "parameters": [
          {
            "description": "File ID returned by add or list Thread files. v1 IDs are bare ULIDs.",
            "example": "01J0000000000000000000000J",
            "in": "path",
            "name": "fileId",
            "required": true,
            "schema": {
              "example": "01J0000000000000000000000J",
              "format": "ulid",
              "pattern": "^[0-7][0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{25}$",
              "type": "string",
              "x-default": "01J0000000000000000000000J"
            }
          },
          {
            "description": "Controls the Content-Disposition response header. Use attachment for downloads or inline for previewable content.",
            "example": "attachment",
            "in": "query",
            "name": "disposition",
            "schema": {
              "default": "attachment",
              "enum": ["attachment", "inline"],
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Download Thread file content",
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Thread file content.",
            "headers": {
              "Cache-Control": {
                "description": "Set to no-store for public API file downloads.",
                "schema": {
                  "type": "string"
                }
              },
              "Content-Disposition": {
                "description": "Download or inline content disposition for the returned file.",
                "schema": {
                  "type": "string"
                }
              },
              "Content-Length": {
                "description": "File size in bytes.",
                "schema": {
                  "minimum": 0,
                  "type": "integer"
                }
              },
              "ETag": {
                "description": "Storage ETag for the returned object.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/files/{fileId}": {
      "delete": {
        "description": "Deletes a pre-Thread uploaded file or a file attached to a public Thread visible to the Access Token caller.",
        "parameters": [
          {
            "description": "File ID returned by add or list Thread files. v1 IDs are bare ULIDs.",
            "example": "01J0000000000000000000000J",
            "in": "path",
            "name": "fileId",
            "required": true,
            "schema": {
              "example": "01J0000000000000000000000J",
              "format": "ulid",
              "pattern": "^[0-7][0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{25}$",
              "type": "string",
              "x-default": "01J0000000000000000000000J"
            }
          }
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Delete a file",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "const": true
                    }
                  },
                  "required": ["ok"],
                  "type": "object"
                }
              }
            },
            "description": "Deleted."
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "get": {
        "description": "Returns public file metadata for a pre-Thread uploaded file or a file attached to a public Thread visible to the Access Token caller.",
        "parameters": [
          {
            "description": "File ID returned by add or list Thread files. v1 IDs are bare ULIDs.",
            "example": "01J0000000000000000000000J",
            "in": "path",
            "name": "fileId",
            "required": true,
            "schema": {
              "example": "01J0000000000000000000000J",
              "format": "ulid",
              "pattern": "^[0-7][0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{25}$",
              "type": "string",
              "x-default": "01J0000000000000000000000J"
            }
          }
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Retrieve file metadata",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "file": {
                    "createdAt": "2026-05-19T00:02:00.000Z",
                    "id": "01J0000000000000000000000J",
                    "mimeType": "text/plain",
                    "name": "brief.txt",
                    "size": 19
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/PublicFileResponse"
                }
              }
            },
            "description": "File metadata."
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/agents/{agentId}/threads": {
      "get": {
        "description": "Returns Threads created by the authenticated Access Token caller.",
        "parameters": [
          {
            "description": "Agent API Endpoint ID from the Agent's API Access panel. v1 IDs are bare ULIDs.",
            "example": "01J00000000000000000000001",
            "in": "path",
            "name": "agentId",
            "required": true,
            "schema": {
              "example": "01J00000000000000000000001",
              "format": "ulid",
              "pattern": "^[0-7][0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{25}$",
              "type": "string",
              "x-default": "01J00000000000000000000001"
            }
          },
          {
            "description": "Filter by archived state: true returns only archived Threads, false only active ones. Omit to return all Threads.",
            "in": "query",
            "name": "archived",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "List Threads for an Agent API Endpoint",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "threads": {
                      "items": {
                        "$ref": "#/components/schemas/ThreadSummary"
                      },
                      "maxItems": 100,
                      "type": "array"
                    }
                  },
                  "required": ["threads"],
                  "type": "object"
                }
              }
            },
            "description": "Thread list."
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "post": {
        "description": "Creates a durable Thread from the latest saved private Agent. Publishing and userId are not required. Input and files are optional. The admitted configuration is frozen for this Thread; subsequent Agent edits affect only new Threads.",
        "parameters": [
          {
            "description": "Agent API Endpoint ID from the Agent's API Access panel. v1 IDs are bare ULIDs.",
            "example": "01J00000000000000000000001",
            "in": "path",
            "name": "agentId",
            "required": true,
            "schema": {
              "example": "01J00000000000000000000001",
              "format": "ulid",
              "pattern": "^[0-7][0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{25}$",
              "type": "string",
              "x-default": "01J00000000000000000000001"
            }
          },
          {
            "description": "Optional key for retry-safe create-thread and send-events calls. Reusing the same key with the same request returns the original response. Reusing the key while the original request is still processing returns 409.",
            "example": "thread-20260528-linear-eng-123",
            "in": "header",
            "name": "Idempotency-Key",
            "schema": {
              "maxLength": 128,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "emptyThread": {
                  "summary": "Create an empty Thread",
                  "value": {}
                },
                "accessTokenWithFile": {
                  "summary": "Access Token with an uploaded file",
                  "value": {
                    "input": {
                      "content": [
                        {
                          "text": "Summarize the attached launch plan and list follow-ups.",
                          "type": "text"
                        }
                      ],
                      "type": "user.message"
                    },
                    "resources": [
                      {
                        "file_id": "01J0000000000000000000000J",
                        "type": "file"
                      }
                    ],
                    "userId": "customer-123"
                  }
                },
                "accessTokenBasic": {
                  "summary": "Access Token",
                  "value": {
                    "input": {
                      "content": [
                        {
                          "text": "Say hello from the API.",
                          "type": "text"
                        }
                      ],
                      "type": "user.message"
                    },
                    "userId": "customer-123"
                  }
                },
                "cattleAgentSameShape": {
                  "summary": "Cattle Agent using the same request shape",
                  "value": {
                    "input": {
                      "content": [
                        {
                          "text": "Run this one-off Public Thread API request.",
                          "type": "text"
                        }
                      ],
                      "type": "user.message"
                    },
                    "userId": "automation"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/CreateThreadRequest"
              }
            }
          },
          "required": false
        },
        "summary": "Create a Thread for an Agent API Endpoint",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateThreadResponse"
                }
              }
            },
            "description": "Created Thread.",
            "headers": {
              "Idempotency-Replayed": {
                "description": "Present as true when this response is replayed from a previous completed request.",
                "schema": {
                  "const": "true",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/threads/{threadId}/archive": {
      "post": {
        "description": "Archives the Thread so it is hidden from default Thread lists.",
        "parameters": [
          {
            "description": "Thread ID returned by create thread. v1 IDs are bare ULIDs.",
            "example": "01J00000000000000000000009",
            "in": "path",
            "name": "threadId",
            "required": true,
            "schema": {
              "example": "01J00000000000000000000009",
              "format": "ulid",
              "pattern": "^[0-7][0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{25}$",
              "type": "string",
              "x-default": "01J00000000000000000000009"
            }
          }
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Archive a Thread",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "ok": {
                      "const": true
                    }
                  },
                  "required": ["ok"],
                  "type": "object"
                }
              }
            },
            "description": "Archived."
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "accessToken": []
    }
  ],
  "servers": [
    {
      "url": "https://cloud.mosoo.ai/api/v2"
    }
  ]
}